requirements.txt
requests
pandas
python-dotenv
numpy
ta
PyYAML
